# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * InsightForge application API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 104857600
     */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface DatasetUploadInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 104857600
     */
  size: number;
  contentType: string;
  /** @minLength 1 */
  objectPath: string;
}

export type DatasetProfile = { [key: string]: unknown };

export interface Dataset {
  id: string;
  name: string;
  fileType: string;
  fileSize: number;
  rows: number;
  columns: number;
  qualityScore: number;
  targetCandidates: string[];
  profile: DatasetProfile;
  createdAt: string;
}

export interface DatasetList {
  datasets: Dataset[];
}

export interface SuccessResponse {
  success: boolean;
}

export interface ProfileResponse { [key: string]: unknown }

export interface StatisticsResponse { [key: string]: unknown }

export type EdaRequestChartType = typeof EdaRequestChartType[keyof typeof EdaRequestChartType];


export const EdaRequestChartType = {
  histogram: 'histogram',
  bar: 'bar',
  scatter: 'scatter',
  box: 'box',
  correlation: 'correlation',
} as const;

export type EdaRequestAggregation = typeof EdaRequestAggregation[keyof typeof EdaRequestAggregation];


export const EdaRequestAggregation = {
  mean: 'mean',
  median: 'median',
  count: 'count',
  sum: 'sum',
} as const;

export interface EdaRequest {
  xColumn?: string;
  yColumn?: string;
  chartType: EdaRequestChartType;
  aggregation?: EdaRequestAggregation;
}

export interface EdaResponse { [key: string]: unknown }

export type CleaningRequestOperationsItemType = typeof CleaningRequestOperationsItemType[keyof typeof CleaningRequestOperationsItemType];


export const CleaningRequestOperationsItemType = {
  remove_duplicates: 'remove_duplicates',
  drop_column: 'drop_column',
  fill_missing: 'fill_missing',
  rename_column: 'rename_column',
  convert_type: 'convert_type',
} as const;

export type CleaningRequestOperationsItemTargetType = typeof CleaningRequestOperationsItemTargetType[keyof typeof CleaningRequestOperationsItemTargetType];


export const CleaningRequestOperationsItemTargetType = {
  number: 'number',
  string: 'string',
  date: 'date',
} as const;

export type CleaningRequestOperationsItem = {
  type: CleaningRequestOperationsItemType;
  column?: string;
  value?: string;
  newName?: string;
  targetType?: CleaningRequestOperationsItemTargetType;
};

export interface CleaningRequest {
  operations: CleaningRequestOperationsItem[];
}

export type TrainRequestProblemType = typeof TrainRequestProblemType[keyof typeof TrainRequestProblemType];


export const TrainRequestProblemType = {
  auto: 'auto',
  classification: 'classification',
  regression: 'regression',
} as const;

export interface TrainRequest {
  datasetId: string;
  /** @minLength 1 */
  targetColumn: string;
  /** @minItems 1 */
  features: string[];
  model: string;
  problemType: TrainRequestProblemType;
  /**
     * @minimum 0.1
     * @maximum 0.5
     */
  testSize?: number;
  /**
     * @minimum 0
     * @maximum 2147483647
     */
  randomSeed?: number;
}

export type ExperimentMetrics = { [key: string]: unknown };

export type ExperimentFeatureImportanceItem = {
  feature: string;
  importance: number;
};

export interface Experiment {
  id: string;
  datasetId: string;
  datasetName: string;
  name: string;
  problemType: string;
  targetColumn: string;
  features: string[];
  model: string;
  metrics: ExperimentMetrics;
  featureImportance: ExperimentFeatureImportanceItem[];
  /** @nullable */
  confusionMatrix?: number[][] | null;
  durationSeconds: number;
  createdAt: string;
}

export interface ExperimentList {
  experiments: Experiment[];
}

export interface ReportRequest {
  datasetId: string;
  experimentId?: string;
}

export type ReportContent = { [key: string]: unknown };

export interface Report {
  id: string;
  datasetId: string;
  title: string;
  content: ReportContent;
  createdAt: string;
}

export interface ReportList {
  reports: Report[];
}

export interface AnalystChatRequest {
  datasetId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
}

export type AnalystChatResponseToolResultsItem = { [key: string]: unknown };

export interface AnalystChatResponse {
  answer: string;
  toolResults: AnalystChatResponseToolResultsItem[];
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};