# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: InsightForge application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai-analyst/chat:
    post:
      operationId: chatWithAnalyst
      tags: [Analyst]
      summary: Answer a dataset question using validated analysis tools
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalystChatRequest" }
      responses:
        "200":
          description: Answer grounded in tool results.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalystChatResponse" }
        "404": { description: Dataset not found. }
        "502": { description: AI provider request failed. }
components:
...
  schemas:
...
    AnalystChatRequest:
      type: object
      required: [datasetId, message]
      properties:
        datasetId: { type: string, format: uuid }
        message: { type: string, minLength: 1, maxLength: 4000 }
    AnalystChatResponse:
      type: object
      required: [answer, toolResults]
      properties:
        answer: { type: string }
        toolResults:
          type: array
          items: { type: object, additionalProperties: true }