# lib/api-zod/src/generated/types/analystChatRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * InsightForge application API
 * OpenAPI spec version: 0.2.0
 */

export interface AnalystChatRequest {
  datasetId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
}